+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/SdkSynchronizer.kt
@@ -65,11 +65,13 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.cancel
+import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.callbackFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
@@ -298,7 +300,7 @@ class SdkSynchronizer private constructor(
 
         val shutdownJob =
             coroutineScope.launch {
-                Twig.debug { "Stopping synchronizer $synchronizerKey…" }
+                Twig.info { "Stopping synchronizer $synchronizerKey…" }
                 processor.stop()
             }
 
@@ -312,6 +314,35 @@ class SdkSynchronizer private constructor(
             instances.remove(synchronizerKey)
         }
     }
+
+    /**
+     * A Flow-providing version of [close] function. This safely closes the Synchronizer together with the related
+     * components.
+     */
+    fun closeFlow(): Flow<Unit> =
+        callbackFlow {
+            val shutdownJob =
+                coroutineScope.launch {
+                    Twig.info { "Stopping synchronizer $synchronizerKey…" }
+                    processor.stop()
+                }
+
+            instances[synchronizerKey] = InstanceState.ShuttingDown(shutdownJob)
+
+            shutdownJob.invokeOnCompletion {
+                coroutineScope.cancel()
+                _status.value = STOPPED
+                Twig.info { "Synchronizer $synchronizerKey stopped" }
+
+                instances.remove(synchronizerKey)
+
+                trySend(Unit)
+            }
+
+            awaitClose {
+                // Nothing to close here
+            }
+        }
 
     override suspend fun getNearestRewindHeight(height: BlockHeight): BlockHeight =
         processor.getNearestRewindHeight(