+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/db/derived/DbDerivedDataRepository.kt
@@ -64,4 +64,8 @@ internal class DbDerivedDataRepository(
     override suspend fun close() {
         derivedDataDb.close()
     }
+
+    override suspend fun isClosed(): Boolean {
+        return derivedDataDb.isClosed()
+    }
 }